pyproject.toml
[project]
name = "email-audit-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "django-rest-framework>=0.1.0",
    "google-genai>=1.20.0",
]
